\name{cart-internal}

%% Documentation should not be exported for the following functions:
%% <LIST>
%% \alias{interpolate}
%% </LIST>

\title{Internal cart functions}
\description{
  Internal cart functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
