\name{cart-package}
\alias{cart-package}
\alias{cart}
\docType{package}
\title{
\packageTitle{cart}
}
\description{
\packageDescription{cart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cart}
\packageIndices{cart}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cart}

Maintainer: \packageMaintainer{cart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
