% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{cartogram}
\alias{cartogram}
\title{Creates a cartogram.}
\usage{
cartogram(spdf, variable = 1, nrows = 2^8, ncols = 2^8)
}
\arguments{
\item{spdf}{\code{\link[sp]{SpatialPolygonsDataFrame}} object for which
to create a cartogram}

\item{variable}{numeric or character indicating the data.frame column in
'spdf' which to use for the density calculations}

\item{nrows}{numeric, number of rows of the sampling grid}

\item{ncols}{numeric, number of columns of the sampling grid}
}
\value{
SpatialPolygonsDataFrame object
}
\description{
This function creates a cartogram from a
\code{\link[sp]{SpatialPolygonsDataFrame}} object using
the indicated data.frame variable.
}
\examples{
library(maps)
library(maptools)
library(sp)
data(usapop)
usapop <- usapop[-c(2, 12, 45), , drop = FALSE]
usmap <- map("state", fill = TRUE, plot = FALSE)
sp <- map2SpatialPolygons(usmap, sub(":.*", "", usmap$names))
rownames(usapop) <- tolower(rownames(usapop))
spdf <- SpatialPolygonsDataFrame(sp, usapop, match.ID = TRUE)
cart <- cartogram(spdf, "population")
plot(cart, axes = TRUE, asp = 1/2, col = "#147f14")
}
\references{
Gastner MT, Newman MEJ (2004) Diffusion-based method for
producing density equalizing maps. Proc. Natl. Acad. Sci. 101:7499-7504
}
\author{
Thomas Zumbrunn \email{thomas@zumbrunn.name}
}
