% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_map}
\alias{get_map}
\title{Get Malaria Atlas Project (MAP) rasters.}
\usage{
get_map(
  raster,
  pop,
  year = NA,
  method = "bilinear",
  quiet = FALSE,
  resample = TRUE
)
}
\arguments{
\item{raster}{Raster}

\item{pop}{Population raster}

\item{year}{Year}

\item{method}{See \code{\link[terra]{rast}} for more information}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar}

\item{resample}{resample to match resolution and extent of pop}
}
\value{
A raster.
}
\description{
Downloads the specified \href{https://malariaatlas.org/}{malaria atlas project} rasters.
}
