% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_prev}
\alias{get_prev}
\title{Get prevalence rasters.}
\usage{
get_prev(prevalence_rasters, year, pop)
}
\arguments{
\item{prevalence_rasters}{List of prevalence rasters. See
\code{\link[malariaAtlas]{listRaster}} for available rasters. Default is
for both falciparum and vivax.}

\item{year}{Year of prevalence outputs}
}
\value{
A List of prevalence rasters.
}
\description{
Downloads the specified \href{https://malariaatlas.org/}{malaria atlas project} prevalence rasters.
}
