% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_spatial_limits}
\alias{get_spatial_limits}
\title{Get spatial limits rasters.}
\usage{
get_spatial_limits(spatial_limits_rasters, pop)
}
\arguments{
\item{spatial_limits_rasters}{List of spatial distribution limits rasters.
Default is for both falciparum and vivax. See
\code{\link[malariaAtlas]{listRaster}} for available rasters.}
}
\value{
A List of spatial limits rasters.
}
\description{
Downloads the specified \href{https://malariaatlas.org/}{malaria atlas project}
  spatial limits of transmission rasters.
}
