% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_vectors}
\alias{get_vectors}
\title{Get vector species.}
\usage{
get_vectors(vector_rasters, year, pop)
}
\arguments{
\item{vector_rasters}{List of vector species rasters.
See \code{\link[malariaAtlas]{listRaster}} for available rasters. Specified
 rasters must have an extent >= to the extent of the country specified.
Please note the different definitions of vector species metrics -
especially relative abundance vs probability of occurrence.}

\item{year}{Year of prevalence outputs}

\item{pop}{Population raster}
}
\value{
A List of vector species rasters.
}
\description{
Downloads the specified \href{https://malariaatlas.org/}{malaria atlas project}
  vector species rasters.
}
