\name{initialize}
\alias{initialize}
\title{Generics for Cartogram class...}
\description{Generics for Cartogram class}
\details{initialize a cartogram; if passed a \link{SpatialPolygonsDataFram},
will compute the cartogram distortion matrix
and store the transfomed \link{SpatialPolygons} in the @cartogram slot}
