% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart.R
\name{interpolate}
\alias{interpolate}
\title{Performs bilinear interpolation.}
\usage{
interpolate(xy, xgrid, ygrid)
}
\arguments{
\item{xy}{numeric matrix of x and y coordinate pairs to be transformed}

\item{xgrid}{numeric matrix of x grid coordinates}

\item{ygrid}{numeric matrix of y grid coordinates}
}
\value{
matrix of x/y transformed coordinates
}
\description{
This function performs bilinear interpolation of pairs of x/y
coordinates based on a matrix of x grid coordinates and a matrix of y
grid coordinates. The algorithm is based on C code by Mark Newman.
}
\author{
Thomas Zumbrunn \email{thomas@zumbrunn.name}
}
