% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{make_urban_rural}
\alias{make_urban_rural}
\title{Get urban rural raster}
\usage{
make_urban_rural(pop, urban_density_threshold)
}
\arguments{
\item{pop}{Population raster (at 1km squared)}

\item{urban_density_threshold}{Threshold for urban classification (people per square km).
Default is 1500 people per square km. Based on urban classification by Eurostat,
detailed in \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8065116/}{Khavari et al}.}
}
\value{
Categorical raster
}
\description{
Recodes the population raster into urban and rural cells, based on a threshold
population density.
}
