% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{pop_available}
\alias{pop_available}
\title{Check if population data are available for a given country and given years}
\usage{
pop_available(iso3c, years)
}
\arguments{
\item{iso3c}{Country iso3c code}

\item{years}{Years}
}
\value{
Boolean vector
}
\description{
Check if population data are available for a given country and given years
}
