% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{raster_available}
\alias{raster_available}
\title{Check if MAP raster data are available for a given country and given years}
\usage{
raster_available(raster_name, years, raster_list)
}
\arguments{
\item{raster_name}{Name of raster}

\item{years}{Years}

\item{raster_list}{Raster metadata from Malaria Atlas Project}
}
\value{
Boolean vector
}
\description{
Check if MAP raster data are available for a given country and given years
}
