% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_cart.R
\name{unpack_cart}
\alias{unpack_cart}
\title{Extract information from rasters}
\usage{
unpack_cart(iso3c_sf, stack)
}
\arguments{
\item{iso3c_sf}{A simple feature shape file to extract for}

\item{stack}{The raster stack from \code{\link{pull_cart}}}
}
\value{
Tibble with list columns of raw extracted values
}
\description{
Extract information from rasters
}
