% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_map}
\alias{get_map}
\title{Get Malaria Atlas Project (MAP) rasters.}
\usage{
get_map(raster, pop, year = NA, method = "bilinear")
}
\arguments{
\item{raster}{Raster}

\item{pop}{Population raster}

\item{year}{Year}

\item{method}{See \code{\link[terra]{rast}} for more information}
}
\value{
A raster.
}
\description{
Downloads the specified \href{https://malariaatlas.org/}{malaria atlas project} rasters.
}
