% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{get_pop}
\alias{get_pop}
\title{Get population raster.}
\usage{
get_pop(iso3c, year)
}
\arguments{
\item{iso3c}{Country iso3c code}

\item{year}{Year}
}
\value{
Population raster
}
\description{
Downloads the unconstrained individual countries 2000-2020 UN adjusted
(1km resolution) from the \href{https://www.worldpop.org/}{WorldPop server}.
}
