% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_cart.R
\name{pull_cart}
\alias{pull_cart}
\title{Pull cartographic information}
\usage{
pull_cart(
  iso3c,
  year,
  vector_rasters = list(funestus = "Anopheles funestus", arabiensis =
    "Anopheles arabiensis Patton, 1905", gambiae = "Anopheles gambiae Giles, 1902"),
  prevalence_rasters = list(pfpr = "Plasmodium falciparum PR2 - 10 version 2020", pvpr
    = "Plasmodium vivax PR1-99 version 2020"),
  spatial_limits_rasters = list(pf_limits = "Plasmodium falciparum Spatial Limits",
    pv_limits = "Plasmodium vivax Spatial Limits"),
  urban_density_threshold = 1500
)
}
\arguments{
\item{iso3c}{Country iso3c code}

\item{year}{Year of prevalence outputs}

\item{vector_rasters}{List of vector species rasters.
See \code{\link[malariaAtlas]{listRaster}} for available rasters. Specified
 rasters must have an extent >= to the extent of the country specified.
Please note the different definitions of vector species metrics -
especially relative abundance vs probability of occurrence.}

\item{prevalence_rasters}{List of prevalence rasters. See
\code{\link[malariaAtlas]{listRaster}} for available rasters. Default is
for both falciparum and vivax.}

\item{spatial_limits_rasters}{List of spatial distribution limits rasters.
Default is for both falciparum and vivax. See
\code{\link[malariaAtlas]{listRaster}} for available rasters.}

\item{urban_density_threshold}{Threshold for urban clasification (people per square km).
Default is 1500 people per square km. Based on urban classification by Eurostat,
detailed in \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8065116/}{Khavari et al}.}
}
\value{
A raster stack
}
\description{
Pull cartographic information
}
